Baca update
Added javadoc and notes to clarify some structure and mark what our
present file names (very difficult to change) correspond with lambert's
similar file names. The GUIView is equivalent to the MainView,
FreeCellModel to FreeCellGame, and FreeCellApp to FreeCellApp.
Significant work on GUIApp is completed, may be near completion but
there are some weird problems/missing parts. Methods akin to
"FreeCellModel.getHomeCell(i)" should be added by my understanding in
the GUIApp (which sounds weird, but thats because it is actually the
FreeCellApp)--to do this the lists holding the cellpanels TopCellPanels
and TableauxPanels would need to be created there as well like the lists
freeCells, homeCells, and tableaux in FreeCellModel (again, my
understanding). Lambert also hints that CellPanel should be an abstract
class (with only the method paintComponent()) extended by two classes
(to be named, I would Assume) TableauxPanel and TopCellPanel since each
display their cards differently. This is my understanding of where the
project needs to go from here.